finalizado

+++ js/registro.js
@@ -1,21 +1,21 @@
-    // Obtener valores del formulario
-    //Estas líneas recuperan los valores introducidos por el usuario en los campos del formulario utilizando los id de los elementos de entrada. El método .value se usa para obtener el valor actual de cada campo de entrada.
-    const nombre = document.getElementById('nombre');
-    const apellidos = document.getElementById('apellidos');
-    const correo = document.getElementById('correo');
-    const telefono = document.getElementById('telefono');
-    const contraseña = document.getElementById('contraseña');
-    const confirmecontraseña = document.getElementById('confirmContraseña');
-    const Registrar = document.getElementById('Registrar');  
+   
+     // Obtener valores del formulario
+     const nombre = document.getElementById('nombre');
+     const apellidos = document.getElementById('apellidos');
+     const correo = document.getElementById('correo');
+     const telefono = document.getElementById('telefono');
+     const contraseña = document.getElementById('contraseña');
+     const confirmecontraseña = document.getElementById('confirmContraseña');
+     const Registrar = document.getElementById('Registrar');  
 
-    Registrar.addEventListener('click', function () {
+        Registrar.addEventListener('click', function () {
         let nombreIngresado = nombre.value.trim();
         let apellidoIngresado = apellidos.value.trim();
         let correoIngresado = correo.value.trim();
         let telefonoIngresado = telefono.value.trim();
         let contraseñaIngresada = contraseña.value.trim();
         let confirmeContraseñaIngresada = confirmecontraseña.value.trim();
-
+            
         // Verificar si las contraseñas coinciden
         if (contraseñaIngresada !== confirmeContraseñaIngresada) {
             alert('Las contraseñas no coinciden.');
@@ -28,12 +28,19 @@
                 telefono: telefonoIngresado,
                 contraseña: contraseñaIngresada
             };
-    
+            if (nombreIngresado === '' && apellidoIngresado === '' && correoIngresado === '' && telefonoIngresado === '' && contraseñaIngresada === '' && confirmeContraseñaIngresada === '') {
+                alert('Complete todos los espacios');
+                return;
+            }
+
+            
+            
+            // Añadir el nuevo usuario a la lista
+            usuarios.push(nuevoUsuario);
+           
             // Obtener la lista de usuarios de localStorage
             let usuarios = JSON.parse(localStorage.getItem('usuarios')) || [];
     
-            // Añadir el nuevo usuario a la lista
-            usuarios.push(nuevoUsuario);
     
             // Guardar la lista actualizada de usuarios en localStorage
             localStorage.setItem('usuarios', JSON.stringify(usuarios));
